style: cast getElValenceforMolecularInorganics return to int
Make the widening from the S_CHAR cValence element to the int return type
explicit. Semantically a no-op (signed char always fits in int), but documents
the conversion for strict static analyzers. Suggested by @djb-rwth.

+++ INCHI-1-SRC/INCHI_BASE/src/strutil.c
@@ -4249,7 +4249,7 @@ int getElValenceforMolecularInorganics(int nPeriodicNum, int charge, int val_num
         return 0;
     }
 
-    return MolecularInorganicsArray[nPeriodicNum > 1 ? nPeriodicNum + 1 : 0].cValence[NEUTRAL_STATE + charge][val_num];
+    return (int) MolecularInorganicsArray[nPeriodicNum > 1 ? nPeriodicNum + 1 : 0].cValence[NEUTRAL_STATE + charge][val_num];
 }
 
 /****************************************************************************
